requirements.txt
# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
requests>=2.25.1
pandas>=1.3.0
python-multipart>=0.0.5

# LLM dependencies
openai>=1.0.0
anthropic>=0.3.0

# Database connectors
mysql-connector-python>=8.0.0
psycopg2-binary>=2.9.0
snowflake-connector-python>=3.0.0
# sqlite3 is included in Python standard library

# Utility dependencies
python-dotenv>=0.19.0
pydantic>=1.8.0
typing-extensions>=4.0.0